修改价格 支持小数

+++ models/admin.go
@@ -24,7 +24,7 @@ type GoodsAdd struct {
 	Gname     string                `json:"gname,omitempty" form:"gname"`
 	Category  string                `json:"type,omitempty" form:"type"`
 	Introduce string                `json:"introduce,omitempty" form:"introduce"`
-	Price     int                   `json:"price,omitempty" form:"price"`
+	Price     float64               `json:"price,omitempty" form:"price"`
 }
 type UpdateGoods struct {
 	GoodsAdd
